# docs/openapi/ferrum.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ferrum",
    "description": "Implementation map of the Ferrum gateway (utoipa dump). Source of truth for GA4GH DRS/WES/TES/TRS/Beacon/htsget is each standard’s published OpenAPI — see docs/GA4GH.md. This file records what this process exposes, including Ferrum-only paths. Not a replacement spec. Not official GA4GH certification. Auth is a deployment choice. Behavioral proof against the published specs is HelixTest.",
    "contact": {
      "name": "Synaptic Four",
      "email": "contact@synapticfour.com"
    },
    "license": {
      "name": "BUSL-1.1",
      "url": "https://github.com/SynapticFour/Ferrum/blob/main/LICENSE"
    },
    "version": "0.3.1"
  },
  "servers": [
    {
      "url": "/",
      "description": "Paths are absolute from the Ferrum gateway origin (default http://127.0.0.1:8080)."
    }
  ],
  "paths": {
...
    "/passports/v1/admin/visa_grants": {
      "get": {
        "tags": [
          "handlers"
        ],
        "summary": "GET /admin/visa_grants - list visa grants (optional ?user_sub= &dataset_id=)",
        "operationId": "admin_list_visa_grants",
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "handlers"
        ],
        "summary": "POST /admin/visa_grants - create a visa grant",
        "operationId": "admin_create_visa_grant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVisaGrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateVisaGrantRequest": {
        "type": "object",
        "required": [
          "user_sub",
          "user_iss",
          "dataset_id",
          "visa_type",
          "value",
          "source"
        ],
        "properties": {
          "conditions": {
            "nullable": true
          },
          "dataset_id": {
            "type": "string"
          },
          "expires_at": {
            "type": "string",
            "description": "ISO8601/RFC3339 datetime string, e.g. \"2025-12-31T23:59:59Z\"",
            "nullable": true
          },
          "source": {
            "type": "string"
          },
          "user_iss": {
            "type": "string"
          },
          "user_sub": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "visa_type": {
            "type": "string"
          }
        }
      },
...
    }
  }
}